Organize all trigger functions into subdirectory
